Refactor booking type date range components
