+++ .git-blame-ignore-revs
@@ -8,3 +8,6 @@
 
 # Run prettier for all files
 3897d1b4465ae8f45c62b2da88cf90033764226c
+
+# Refactor frontend folder structure
+83b25dcb0e8f2c361261a4757ff7ae392713c812
